requirements.txt
torch>=2.0.0
transformers>=4.36.0
PyPDF2>=3.0.0
sentencepiece
textstat
sentence-transformers
scikit-learn
spacy
streamlit
networkx
groq
python-dotenv
